refactor: share UI helper functions
